Add files via upload
A preliminary software manual is provided in the spallmap_manual17.pdf file.  The manual include instructions for installation, a description of main parameters used in the input file and a description of the output, including some instructions (line commands in UNIX and UNIX scripts) on how to extrat the data from the output file.